Move all messenger code to chat_service/services/messenger
